fix: add field resolver for festID in transaction type
