+++ api/src/apollo/index.ts
@@ -3,7 +3,7 @@ export * from './User';
 export * from './Org';
 export * from './Event';
 export * from './Team';
-export * from './Transactions';
+export * from './Transaction';
 export * from './EventRegistration';
 export * from './Story';
 export * from './Location';
